Managed comments

+++ events_api/lib/events_api_web/controllers/comment_controller.ex
@@ -2,42 +2,49 @@ defmodule EventsApiWeb.CommentController do
   use EventsApiWeb, :controller
 
   alias EventsApi.Comments
+  alias EventsApi.Events
   alias EventsApi.Comments.Comment
+  alias EventsApiWeb.Plugs
+
+  plug Plugs.RequireAuth when action in [:create, :delete]
 
   action_fallback EventsApiWeb.FallbackController
 
-  def index(conn, _params) do
-    comments = Comments.list_comments()
-    render(conn, "index.json", comments: comments)
-  end
-
   def create(conn, %{"comment" => comment_params}) do
-    with {:ok, %Comment{} = comment} <- Comments.create_comment(comment_params) do
-      conn
-      |> put_status(:created)
-      |> put_resp_header("location", Routes.comment_path(conn, :show, comment))
-      |> render("show.json", comment: comment)
+    user = conn.assigns[:current_user]
+    event = comment_params
+            |> Map.get("event_id")
+            |> Events.get_event!
+            |> Events.load_event
+
+    invited? = Enum.reduce(event.invites, false, &(&2 || &1.email == user.email))
+    if user.id != event.owner && !invited? do
+      conn  
+      |> send_resp(:unprocessable_entity, Jason.encode!(%{"error" => "Not allowed"}))
+    else 
+      with {:ok, %Comment{} = comment} <- Comments.create_comment(Map.put(comment_params, "user_id", user.id)) do
+        conn
+        |> put_status(:created)
+        |> put_resp_header("location", Routes.comment_path(conn, :show, comment))
+        |> send_resp(:ok, "{}")
+      end
     end
   end
 
-  def show(conn, %{"id" => id}) do
-    comment = Comments.get_comment!(id)
-    render(conn, "show.json", comment: comment)
-  end
-
-  def update(conn, %{"id" => id, "comment" => comment_params}) do
-    comment = Comments.get_comment!(id)
-
-    with {:ok, %Comment{} = comment} <- Comments.update_comment(comment, comment_params) do
-      render(conn, "show.json", comment: comment)
-    end
-  end
 
   def delete(conn, %{"id" => id}) do
     comment = Comments.get_comment!(id)
-
-    with {:ok, %Comment{}} <- Comments.delete_comment(comment) do
-      send_resp(conn, :no_content, "")
+    user = conn.assigns[:current_user]
+    event = Events.get_event!(comment.event_id)
+    if user.id != event.owner && user.id != comment.user_id do
+      send_resp(conn, :unprocessable_entity, "Not Allowed")
+    else 
+      event = comment.event_id
+              |> Events.get_event!
+              |> Events.load_event
+      with {:ok, %Comment{}} <- Comments.delete_comment(comment) do
+        send_resp(conn, :no_content, "")
+      end
     end
   end
 end
